Quiz Prep/class-1/pyproject.toml
[project]
name = "class-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "openai-agents" # Replace with actual package name if different
]
